# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Application",
        "title": "ActionDetector API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4040",
    "paths": {
        "/api/actions": {
            "get": {
                "description": "Get a list of all previously predicted actions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Get all actions",
                "responses": {
                    "200": {
                        "description": "Returns the list of actions",
                        "schema": {
                            "$ref": "#/definitions/handler.getActionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete all previously predicted actions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Delete all actions",
                "responses": {
                    "200": {
                        "description": "Successfully deleted all actions",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/files": {
            "post": {
                "description": "Upload a video file with size limit of 50MB",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "video"
                ],
                "summary": "Upload a video file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Video file to upload (max 50MB)",
                        "name": "video",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the file link",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid file size or type",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/predict": {
            "post": {
                "description": "Upload a video URL and get predicted actions in Uzbek language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prediction"
                ],
                "summary": "Predict actions from video",
                "parameters": [
                    {
                        "description": "Request body with video URL",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.predictActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns predicted actions in Uzbek",
                        "schema": {
                            "$ref": "#/definitions/handler.pretictActionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error or prediction service error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.getActionsResponse": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Action"
                    }
                }
            }
        },
        "handler.predictActionRequest": {
            "description": "Request containing the URL of the video to analyze",
            "type": "object",
            "required": [
                "video_url"
            ],
            "properties": {
                "video_url": {
                    "description": "@Description URL of the video to analyze",
                    "type": "string"
                }
            }
        },
        "handler.pretictActionResponse": {
            "description": "Response containing predicted actions in Uzbek language",
            "type": "object",
            "properties": {
                "predictions": {
                    "description": "@Description List of predicted actions in Uzbek language",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.Action": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "video_id": {
                    "type": "string"
                }
            }
        },
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}